QUAL-216 Separates elastic component registry
