+++ src/Esfa.Vacancy.Infrastructure/Ioc/InfrastructureRegistry.cs
@@ -8,11 +8,9 @@ using Esfa.Vacancy.Domain.Interfaces;
 using Esfa.Vacancy.Infrastructure.Caching;
 using Esfa.Vacancy.Infrastructure.Services;
 using Esfa.Vacancy.Infrastructure.Settings;
-using SFA.DAS.Elastic;
 using SFA.DAS.NLog.Logger;
-using SFA.DAS.VacancyServices.Search;
 
-namespace Esfa.Vacancy.Infrastructure
+namespace Esfa.Vacancy.Infrastructure.Ioc
 {
     public sealed class InfrastructureRegistry : StructureMap.Registry
     {
@@ -40,31 +38,8 @@ namespace Esfa.Vacancy.Infrastructure
                 .Ctor<ITrainingDetailService>().Is<TrainingDetailService>();
 
             RegisterCreateApprenticeshipService();
-
-            RegisterElasticsearchClient();
         }
 
-        private void RegisterElasticsearchClient()
-        {
-            var username = _provideSettings.GetNullableSetting(ApplicationSettingKeys.ElasticUsernameKey);
-            var password = _provideSettings.GetNullableSetting(ApplicationSettingKeys.ElasticPasswordKey);
-            var indexName = _provideSettings.GetSetting(ApplicationSettingKeys.ApprenticeshipIndexAliasKey);
-
-            ElasticClientConfiguration elasticConfig;
-#if DEBUG
-            var hostUrl = _provideSettings.GetSetting(ApplicationSettingKeys.VacancySearchUrlKey);
-
-            elasticConfig = string.IsNullOrEmpty(username) || string.IsNullOrEmpty(password) 
-                ? new ElasticClientConfiguration(new Uri(hostUrl))
-                : new ElasticClientConfiguration(new Uri(hostUrl), username, password);
-#else
-            var cloudId = _provideSettings.GetSetting(ApplicationSettingKeys.ElasticCloudIdKey);
-            elasticConfig = new ElasticClientConfiguration(cloudId, username, password);
-#endif
-            For<ElasticClientConfiguration>().Use(elasticConfig);
-            For<IElasticClientFactory>().Use<ElasticClientFactory>();
-            For<IApprenticeshipSearchClient>().Use<ApprenticeshipSearchClient>().Ctor<string>("indexName").Is(indexName);
-        }
 
         private IDictionary<string, object> GetProperties()
         {
